refactor(studio): extract usePersistedValue, dedupe origin
Both flagged reviewing the explorer-parity branch:

- Add a generic usePersistedValue<T>(key, fallback) to
  browser-storage.ts (a scalar over usePersistedList) and route the
  SQL split-view and the tabs' active-id through it, replacing two
  hand-rolled one-element-list wrappers.
- Hoist a canonical resolveOrigin(explicit?) into lib/internal.ts;
  app.tsx's resolveBaseUrl and the connect-agent card now share it,
  so the dev-origin fallback lives in one place instead of drifting
  across two files.

Behaviour-preserving; typecheck/eslint/pure-logic tests green.
